fixed RPM errors in dv

+++ Core/Inc/CAN_utils.h
@@ -229,7 +229,7 @@ void can_bus_send_bms_precharge_state(uint8_t precharge_state, CAN_HandleTypeDef
 
 /* Autonomous bus functions */
 
-void can_send_vcu_rpm(CAN_HandleTypeDef *hcan, uint16_t rpm);
+void can_send_vcu_rpm(CAN_HandleTypeDef *hcan, long rpm);
 
 void can_send_autonomous_HV_signal(CAN_HandleTypeDef *hcan, uint8_t hv_state);
